perf: Add support for creating events for offline CPU
Perf framework APIs for registering and unregistering events
does not work when the CPU is offline. Registering of an
event on a CPU that is offline could start counting when
the CPU comes up.

So introduced new state such that, events can created on an
offline cpu, but is not usable yet. As soon as the CPU comes
up back online,the event is completely created and starts counting.

Signed-off-by: ethan-halsall <ethan.halsall@gmail.com>