cpu_input_boost: implement and hook general boosting
The general boosting API is designed to be more battery efficient when
boosting in miscellaneous drivers than forcing all boosting upon the
input boosting interface.

This commit initializes and hooks general boosting into mdss frame
readiness, however this API is available throughout the kernel, and can
be utilized in a variety of cases.

Signed-off-by: ethan-halsall <ethan.halsall@gmail.com>