defconfig: Decrease max stune boost to 15
Sometimes the little CPU cluster will get stuck at its max frequency
(1766400 KHz) for a while after a max stune boost occurs when the boost
value is 100. This is due to a bug found by RenderBroken in CAF sdm845
scheduler code where stune-boosted top-app tasks will run on the little
cluster instead of the big cluster when the big cluster is NOT
overutilized, but using a boost margin calculated for the big cluster.
As a result, the little cluster is constantly boosted to 1766400 KHz.

This bug affects both power usage and performance when it occurs:
 - Power usage is higher due to the little cluster getting stuck at such
 a high frequency, as well as being overutilized in some cases
 - Performance is lower due to problematic tasks running on the little
 cluster when they should be running on the big cluster

There is no known fix at the moment, so work around the bug by using a
lower boost value of 15 to be safe. cpu_input_boost will still boost the
CPU frequencies to their respective maximums.

Test: freq no longer gets locked to 1766400 KHz for several minutes on
the little cluster after turning off the screen with max stune boost
Reported-by: RenderBroken <zkennedy87@gmail.com>
Signed-off-by: Danny Lin <danny@kdrag0n.dev>
Signed-off-by: ethan-halsall <ethan.halsall@gmail.com>