cpu_input_boost: Add API to check whether to boost on frame commit
There is a trend of kicking boost whenever a new frame is committed to
the display stack. This has the unfortunate side effect of overly
boosting during workloads that constantly update the screen but are not
very CPU-intensive, with video playback being a key example. Over time,
such excessive boosting can add up to a noticeable hit on battery life.

To mitigate this flaw, add a configurable timeout that restricts frame
boosting to only apply within the specified amount of time since the
last input event was received, as input events are indicative of
interaction.

The default timeout of 3250 ms has been empirically determined by
performing multiple fling interactions at various speeds while timing
the duration of the fling. The maximum fling time was observed to be
approximately 2850 ms, so our frame boost timeout of 3250 ms offers some
headroom for more intensive applications such as games.

It is the responsibility of boost kick callers to use the new function
to check whether they should kick frame boost.

Boost timeout values:
  - negative: always boost
  - zero: never boost
  - positive: boost within timeout of X ms

Signed-off-by: Danny Lin <danny@kdrag0n.dev>
Signed-off-by: ethan-halsall <ethan.halsall@gmail.com>